refactor(experiment): split run record parameter modules
