+++ rust/ommx/src/experiment/commit.rs
@@ -1,6 +1,7 @@
 //! Sealing an experiment session into an immutable OMMX Artifact.
 
-use super::model::{ParameterValue, RecordRef, UnsealedExperimentState};
+use super::parameter::ParameterTable;
+use super::{RecordRef, UnsealedExperimentState};
 use super::{
     ANN_ARTIFACT_KIND, ANN_EXPERIMENT_NAME, ANN_EXPERIMENT_SCHEMA, ANN_EXPERIMENT_STATUS,
     ANN_LAYER, ARTIFACT_KIND_EXPERIMENT, EXPERIMENT_INDEX_MEDIA_TYPE, EXPERIMENT_SCHEMA_V1,
@@ -14,7 +15,7 @@ use crate::artifact::{media_types, ImageRef, LocalArtifact};
 use anyhow::Result;
 use oci_spec::image::MediaType;
 use serde::Serialize;
-use std::collections::{BTreeMap, HashMap};
+use std::collections::HashMap;
 
 impl<'reg> UnsealedExperimentState<'reg> {
     /// Consume the unsealed experiment state and commit it as one
@@ -166,114 +167,6 @@ fn run_attributes_json(state: &UnsealedExperimentState<'_>) -> RunAttributes {
 
 fn run_parameters_json(state: &UnsealedExperimentState<'_>) -> Result<ParameterTable> {
     ParameterTable::from_runs(state)
-}
-
-#[derive(Serialize)]
-struct ParameterTable {
-    columns: BTreeMap<String, ParameterColumn>,
-}
-
-impl ParameterTable {
-    fn from_runs(state: &UnsealedExperimentState<'_>) -> Result<Self> {
-        let mut columns = BTreeMap::new();
-        for run in &state.runs {
-            for (name, value) in &run.parameters {
-                columns
-                    .entry(name.clone())
-                    .or_insert_with(|| ParameterColumn::from_value(value))
-                    .insert(name, run.run_id, value)?;
-            }
-        }
-        Ok(Self { columns })
-    }
-}
-
-#[derive(Serialize)]
-#[serde(tag = "type", content = "values")]
-enum ParameterColumn {
-    #[serde(rename = "bool")]
-    Bool(BTreeMap<u64, bool>),
-    #[serde(rename = "int64")]
-    Int(BTreeMap<u64, i64>),
-    #[serde(rename = "float64")]
-    Float(BTreeMap<u64, f64>),
-    #[serde(rename = "string")]
-    String(BTreeMap<u64, String>),
-}
-
-impl ParameterColumn {
-    fn from_value(value: &ParameterValue) -> Self {
-        match value {
-            ParameterValue::Bool(_) => Self::Bool(BTreeMap::new()),
-            ParameterValue::Int(_) => Self::Int(BTreeMap::new()),
-            ParameterValue::Float(_) => Self::Float(BTreeMap::new()),
-            ParameterValue::String(_) => Self::String(BTreeMap::new()),
-        }
-    }
-
-    fn insert(&mut self, name: &str, run_id: u64, value: &ParameterValue) -> Result<()> {
-        match (self, value) {
-            (Self::Bool(values), ParameterValue::Bool(value)) => {
-                values.insert(run_id, *value);
-                Ok(())
-            }
-            (Self::Int(values), ParameterValue::Int(value)) => {
-                values.insert(run_id, *value);
-                Ok(())
-            }
-            (column @ Self::Int(_), ParameterValue::Float(value)) => {
-                let mut values = match std::mem::replace(column, Self::Float(BTreeMap::new())) {
-                    Self::Int(values) => values
-                        .into_iter()
-                        .map(|(run_id, value)| (run_id, value as f64))
-                        .collect::<BTreeMap<_, _>>(),
-                    _ => unreachable!(),
-                };
-                values.insert(run_id, *value);
-                *column = Self::Float(values);
-                Ok(())
-            }
-            (Self::Float(values), ParameterValue::Int(value)) => {
-                values.insert(run_id, *value as f64);
-                Ok(())
-            }
-            (Self::Float(values), ParameterValue::Float(value)) => {
-                values.insert(run_id, *value);
-                Ok(())
-            }
-            (Self::String(values), ParameterValue::String(value)) => {
-                values.insert(run_id, value.clone());
-                Ok(())
-            }
-            (column, value) => {
-                crate::bail!(
-                    "Run parameter `{name}` has mixed column types: existing {}, incoming {}",
-                    column.type_name(),
-                    value.type_name()
-                )
-            }
-        }
-    }
-
-    fn type_name(&self) -> &'static str {
-        match self {
-            Self::Bool(_) => "bool",
-            Self::Int(_) => "int64",
-            Self::Float(_) => "float64",
-            Self::String(_) => "string",
-        }
-    }
-}
-
-impl ParameterValue {
-    fn type_name(&self) -> &'static str {
-        match self {
-            Self::Bool(_) => "bool",
-            Self::Int(_) => "int64",
-            Self::Float(_) => "float64",
-            Self::String(_) => "string",
-        }
-    }
 }
 
 #[derive(Serialize)]
